Refactor vectorization
