+++ burn-wgpu/src/codegen/dialect/gpu/mod.rs
@@ -2,8 +2,10 @@ mod body;
 mod operation;
 mod shader;
 mod variable;
+mod vectorization;
 
 pub(crate) use body::*;
 pub(crate) use operation::*;
 pub(crate) use shader::*;
 pub(crate) use variable::*;
+pub(crate) use vectorization::*;
