+++ burn-wgpu/src/codegen/dialect/wgsl/mod.rs
@@ -5,6 +5,8 @@ mod extension;
 mod operations;
 mod shader;
 
+pub(crate) use base::*;
+pub(crate) use body::*;
 pub(crate) use compiler::*;
 pub(crate) use extension::*;
 pub(crate) use operations::*;
